Ignore automatically generated files
I noticed when building and testing the dist, that a `MYMETA.json.lock`
file was produced and thought it might be a good idea to ignore this
file so that it doesn't inadvertently get added to the repo.

Also, when running `cover` the `cover_db/` directory is created and this
doesn't need to be in the repo either, hence it's also added to the
ignore list.